refactor(ui): restructure components and update imports for improved organization.
