+++ src/app/globals.css
@@ -191,21 +191,18 @@
     width: 100%;
     margin-left: auto;
     margin-right: auto;
-    padding-left: 1rem;
-    padding-right: 1rem;
+    padding: 1rem;
   }
 
   @media (min-width: 768px) {
     .wrapper {
-      padding-left: 1.5rem;
-      padding-right: 1.5rem;
+      padding: 1.5rem;
     }
   }
 
   @media (min-width: 1024px) {
     .wrapper {
-      padding-left: 2rem;
-      padding-right: 2rem;
+      padding: 2rem;
     }
   }
