+++ src/app/page.tsx
@@ -1,27 +1,27 @@
-import AnimatedHero from "@/components/AnimatedHero";
-import { AnimatedTestimonialSection } from "@/components/AnimatedTestimonialSection";
-import TrustedByExpertsSection from "@/components/TrustedByExpertsSection";
-import CallToActionPage from "@/components/CallToActionPage";
-import { TrainingPartner } from "@/components/TrainingPartner";
-import { Sponsors } from "@/components/Sponsors";
-import HowUProWorks from "@/components/HowUProWorks";
+import HomeHero from "@/components/home/HomeHero";
+import { Comments } from "@/components/home/Comments";
+import TrustedByExpertsSection from "@/components/home/TrustedByExpertsSection";
+import CallToAction from "@/components/home/CallToAction";
+import { TrainingPartner } from "@/components/home/TrainingPartner";
+import { Sponsors } from "@/components/home/Sponsors";
+import HowUProWorks from "@/components/home/HowUProWorks";
 
 // Server Component - rendered on the server
 export default function Home() {
   return (
     <main
       id="main-content"
-      className="bg-[#020d02] flex flex-col gap-16"
+      className="bg-black flex flex-col gap-16"
       role="main"
       aria-label="U-Pro Soccer homepage"
     >
-      <AnimatedHero />
+      <HomeHero />
       {/* <AnimatedHowUProWorks /> */}
       <HowUProWorks />
       <TrustedByExpertsSection />
       <TrainingPartner />
-      <CallToActionPage />
-      <AnimatedTestimonialSection />
+      <CallToAction />
+      <Comments />
       <Sponsors />
     </main>
   );
