+++ src/components/home/CardDetails.tsx
